+++ src/Frontend/Dfe.Complete/Startup.cs
@@ -1,22 +1,22 @@
 using Azure.Storage.Blobs;
 using Dfe.Complete.Application.Common.Mappers;
-using Dfe.Complete.Authorization;
 using Dfe.Complete.Configuration;
+using Dfe.Complete.Infrastructure;
+using Dfe.Complete.Infrastructure.Security.Authorization;
 using Dfe.Complete.Security;
+using Dfe.Complete.Services;
 using Dfe.Complete.StartupConfiguration;
 using DfE.CoreLibs.Security.Authorization;
 using GovUk.Frontend.AspNetCore;
 using Microsoft.AspNetCore.Authentication.Cookies;
-using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.DataProtection;
 using Microsoft.AspNetCore.HttpOverrides;
 using Microsoft.FeatureManagement;
 using Microsoft.Identity.Web;
 using Microsoft.Identity.Web.UI;
-using System.Security.Claims;
-using Dfe.Complete.Infrastructure;
-using Dfe.Complete.Infrastructure.Security.Authorization;
-using Dfe.Complete.Services;
+using DfE.CoreLibs.Security.Cypress;
+using Microsoft.AspNetCore.Authentication.OpenIdConnect;
+using System.Configuration;
 
 namespace Dfe.Complete;
 
@@ -78,13 +78,18 @@ public class Startup
         services.AddHttpContextAccessor();
 
         services.AddApplicationAuthorization(Configuration);
-        
-        services.AddMicrosoftIdentityWebAppAuthentication(Configuration);
+
+        var authenticationBuilder = services.AddAuthentication(options =>
+        {
+            options.DefaultScheme = "MultiAuth";
+            options.DefaultChallengeScheme = OpenIdConnectDefaults.AuthenticationScheme;
+        }).AddCypressMultiAuthentication();
+
+        authenticationBuilder.AddMicrosoftIdentityWebApp(Configuration);
+
         ConfigureCookies(services);
 
         services.AddApplicationInsightsTelemetry();
-
-        services.AddSingleton<IAuthorizationHandler, HeaderRequirementHandler>();
 
         System.Text.Encoding.RegisterProvider(System.Text.CodePagesEncodingProvider.Instance);